refactor(renderer): split html renderer modules
